Tychools merging segments
We are merging the segments in a single elf binary now.
I had to fix a small bug in the phdrs headers (wrong offset).
I added a command to dump the page tables.

Next: We might have to add the sections too, not sure yet how to deal
with that, it requires fixing shstrab which will be a pain.