Implementing support for traps and cores
Each domain is only authorized to handle certain traps itself and run on
specific cores.
For the moment, we ignore the cores but I need the traps to work for
testing and VMs exploration.

I implemented two new monitor calls to set traps and cores per domain
BEFORE it is sealed.
The code checks that the traps and cores are a (possibly equal) subset
of the parent's ones.

This commit also adds a new UpdateTrap update per core to request
a change of the set_exception_bitmap on the VCPU.

The next commit should update the sdklibrary to set the cores and traps.