Core filtering for domains
Domains now have a core_map that specifies the physical cores they are
allowed to run on.

The core_map is checked whenever we perform a switch.
:warn We do not check the core on exception routing as by construction
it can only return to a manager core that was allowed. If we decide to
allow dynamic reconfiguration of cores, we will need to handle that
case.