S1: smarter memory partitioning
The first stage used to reserve one whole memory region for itself. On
QEMU this was not posing issues as the memory regions are small, but on
real hardware 60 out of 64GB were reserved for Tyche.

This commit allows reserving a portion of a region instead of the whole
region, solving the above issue.