A skeleton of network adapter for redis
The code opens a tcp server, accepts connections and serves them on
a separate thread. The thread spawns a consumer thread and forwards
messages on a shared memory channel to the consumer thread.
The consumer thread prints the message on the screen.
The consumer thread will be replaced by redis (and redis will have
a default channel from the begining while threads serving different
clients will synchronize to write their messages to the channel I guess)