Introducing engine try-lock-or-update
There was a deadlock if a tlb shootdown was sent in-between a vmcall and
the handle_exit attempting to lock the capa-engine.
The thread sending the TLBshootdown signal has a lock on the engine.
The thread receiving it  does not see the signal as it is in the monitor
just about to call a function attempting to lock the engine.
The result is a deadlock.
To solve this, we introduce a try lock around the engine that, if the
locking fails, processes core updates.