Add support for decoding attestations in tychools
For now the new command simply prints out the decoded attestation, but
does nothing else special.
This exists in parallel with the current attestation mechanism, but the
end vision is to replace the old one by this new implementation in the
future, once we encompass enough features (e.g. hashes).