:rocket: New encr-redis benchmark + core-fix
1) The encr-redis benchmark combines ping-pong and redis:

The untrusted code communicates with an encr domain that `decrypts` (for
the moment does not do anything) a request, forwards it to a redis
enclave via a pipe, reads the response, `encrypts` it (for the moment
does not do anything), and forwards it to the untrusted code.

2) Fixing the core map to core id:

There was a bug in the core map to core id translation.