Fixing BUG in sdktyche/loader
We made the wrong assumption that all segment starts were page aligned.
This lead to funky situations where we copied the content at the wrong
address, leading to e.g., global variable with init values being
0 (this happened with some globals in ssl).

I took this as an opportunity to fix the size computation too that was
potentially incorrect.
To get the real size we need to align_up(start + size)
- align_down(start).