Fixing merge bug
Merging Neelu's code and the refactoring introduced a bug that replaced
the revoke_aliased with a simple revoke leading to weird behaviors when
running td1.

I fixed this by separating the two again (as was done originally) and
adding an ifdef in the driver code for linux to only call revoke on
riscv (instead of revoke_alias).