Preemption when a running domain is revoked.
This commit implements the preemption of a running domain.
The initiating core will send IPIs to the ones that have scheduled the
domain.
These cores will exit and process the IPI, notify they exited on
a barrier, and block on the domain they need to return too (the manager
of the one being preempted).

The initiating revocation domain uses its lock on the capa engine to
revoke the domain. It then notifies the waiting cores that the
revocation completed.

I ran through a few issues that required to adapt the code:
1. The driver was attempting to use revoked capabilities (e.g., read gp)
   and had to be adapted. I also had to modify the monitor to return
   errors in such cases rather than panic.
2. This led to handling error reporting better in the linux drivers.
   I have a skeleton for more general platform independent error codes.,
   including domain revoked.
3. As this is now a valid scenario, I replaced some "ERROR" prints with
   "DEBUG" in linux to avoid poluting the screen.

Bonus:
I simplified the `simple-enclave` example to make it really simple and
remove the old attestation code (which was migrated into its own folder
in `C/libraries/skdtyche/example`).