Finally fixing Efer bug! :bug: :warning:
We have been carrying around an efer bug for the longest time.
Our efer in tyche was set to 0xd00, which means no sysretq support.
All was fine with just dom0 or sdk-based tds because we enable the
save/restore efer on entries/exits for the host and guest.
However, it was the case that sometimes we would have a bad efer and
this is where the weird msr write in platform.rs upon switches came from
(an odd bug fix that seemed to mostly solve the issue).

Now however, as we do run gramine enclaves inside td1, the bug became
apparent.
KVM has this weird behavior of not doing a save/restore of efer if the
host (td0) and the guest (td1) have the same value.
This resulted in a weird state when running td2 (gramine) where we'd
loose the SCE EFER bit (bit index 0), leading to a double fault while
running td1.

The fix was rather simple after that, I just added the bit to tyche
itself.
Note however that this is an issue we should consider more broadly as we
might need to do extra work to fully virtualize these MSRs and take into
account the KVM case.