Add support for CPUID emulation in Tyche
CPUID is one of the most frequent cause of exits, and crucially triggers
many more exits when emulated by KVM in TD0. This patch adds support for
emulating CPUID in Tyche by caching some of the CPUID entries (like KVM
does).
The CPUID entries are supplied by the domain's manager, as it is just a
cache if a CPUID entry is not present Tyche will still transfer control
to the domain's manager which will handle the emulation itself. This
mechanism mimics the KVM cpuid implementation, which accepts a list of
entries from the user-space VMM.