:star: Confidential linux vms + bug fix :bug:
This commit introduces a new linux kconfig for confidential VMs.
After talking with Charly, we figured it was just better to create
a seperate configuration and compilation pipeline as this entails
switching on swiotlbs and cmas.
I adapted the justfile to generate the new kernel under
`builds/linux-x86-nested-confidential` and the setup lab rule to install the
corresponding kernel under `/tyche/vms/bzImageConfidential`.

The :bug: fix is located in the capability region.
The logic for removing a region was flawed and ended up removing regions
that partially overlaped (at the begining but not the end) with the
removed range. We should inspect this commit when merging to ensure the
new logic is correct.

I also disable the ramfs for nested vms so that it picks up the nested
disk.

Bonus: I added a cpuid-enclave example in the sdk in case we need to
further debug/study the behavior of cpuid.

Switching:
Adding a differentiating value to switch to differentiate between
interrupt return and synchronous return.