Remove vmclear calls
It turns out we don't migrate vCPUs once created, this makes it possible
to optimize the frame switch by not calling vmclear. The only place
vmclear is needed is after vCPU creation, as creation can be done from
another core.

It turns out our VMX abstraction was a bit off, the launched state must
be tracked at the level of the vCPU not the VMX frame as we were doing.
This commit bubbles-up the gestion of the launched bit to Tyche (which
creates the vCPU abstraction).