Adding debugging + transition binaries
Throughout the development of tyche we used some small binaries for
debugging and one to allow a confidential VM to revoke a parent's access
to its memory. I am saving these small C files in the repo in case we
need to use them.
There is also a collection of small test functions for futex that was
used while developping support for gramine.

I also removed nested.raw as a qcow2 candidate