First commit of cpuid support
We have run into this issue before, the core id mapping done by the
hardware and by linux might be different. For the moment we fixed it by
using remap core inside the monitor.
However, I discovered that it is possible for processors to have
non-continuous ids, e.g., 0,1,2,3,6,7,8,9.
As a result, we need a better solution to support any hardware in the
long run.

The way cpuid is assigned is as follows:
The hardware lists the processor cores in the APIC.
The hardware cpu ids are not fully ordered and usually look something like
this on hardware with hyperthreading: 0, 2, 4, 6, 1, 3, 5, 7 on (in virtual
environments, the ids are ordered). The list might  even have missing
numbers, e.g., 0, 2, 6, 8, 1, 3, 7, 9.
Linux looks at the order in which they appear in the apic and numbers
them with logical IDs in order of appearance.

What this commit does is keep the order, remove the assumption that
numbers are continuous, and passes an smp_map to the next stage where
indices are real cpuid (apic ID) and values logical ids (order of
appearance).
In the next commit, we will use this information to replace cpuid()
occurences in the monitor with logical ids, so that we speak the same
language as dom0 and see a continuous id space.