Adding explicit types for logical and physical cpuids
This commit adds explicit types LogicalID and physicalID to distinguish
the two. It does not relaly make a massive difference, just makes the
assumptions about which kind of ID is expected explicit in the code.
As they are declared as structs, typecast to usize or u32 has to be
explicitely called.
I changed the interfaces in the code too, but not in the capa engine
yet.

We will need to remove the remap_cores to test on baremetal linux

I ported the riscv code despite such differences between logical and
physical IDs not making sense necessarily on riscv platforms.

:warning: I tested in a virtual environment on x86, but not inside
riscv. We should do that