Adding support for devices in stage 1 and 2
We discover new things everyday:
The device address space can go beyond the RAM physical address space.
As a result, it is necessary for stage 1 to correctly parse the BARs
from the hardware configuration space and determine the phsyical
addresses that correspond to devices.
These addresses must be passed to stage 2 such that device capabilities
(a region capa with RegionKind::Device) can be created and the
corresponding region mapped into the domain's EPTs.

Limitation: for the moment I disable alias and carve on device regions.
We will re-enable such operations once we decide how to expose them in
a principled way.