Implementing support for devices correctly.
We need to parse the BARs for each devices and register corresponding
capabilities in tyche. The devices should be mapped as unchacheable to
be correctly accessed.
I added an "uncacheable" attribute inside the access rights. It counts
towards the count in the tracker preventing merging of cachable and
uncachable regions.
I also updated the tests as now regions print this information out as
well.