IOMMU working on baremetal MSR
I modeled the iommus as devices for the moment to make them mapped as
uncachable when exposed directly to TD0 linux.
Obviously, the end goal is to remove them from linux and paravirtualize
them by translating whatever linux attempts to do into capability
operations.
Worth noting, it seems that mapping them as cachable into a non-root
domain triggers an INIT signal upon access (even read).

Another note, I discovered that iommus must be enabled in order for
Linux to accept to remain in x2apic mode (MSR based operations) as the
IOMMU is relied upon to provide interrupt remapping.
Linux has a tendency to fall back to xapic otherwise.

Finally, I also discovered that BIOS or UEFI can flip a bit to tell
linux to favor xapic rather than x2apic. A linux kernel command line
argument `intremap=no_x2apic_optout` can be used to ignore that bit.
We should investigate what is going on with our UEFI support, it is the
second time we have an issue (the first was the USB controller in the
previous commit, where nicely transfering the ownership of the USB from
BIOS to Linux was stalling).