Finishing support for chain-1 interrupt filtering and routing
:warning:

This requires to set acknowledge interrupt on exit as we should
distinguish betwene different external interrupts and thus need to
capture the vector.
This version of the code is maybe not correct with regards to the state
of the x2apic. we should discuss whether or not to virtualize it.